+++ markov_chain.h
@@ -15,7 +15,7 @@ typedef struct discrete_chain discrete_chain_t;
  * The chain is created based on a file sent in, which is
  * opened and parsed. A suitable format may be the JSON-format.
  */
-discrete_chain_t *chain_create(double *matrix, int num_dim, int start);
+discrete_chain_t *chain_create(float *matrix, int num_dim, int start);
 
 /*
  * The chain is destroyed, freeing all allocations from memory.
